# Improve performance of chacha20-poly1305 on AVX and AVX2.