@toddouska 2018-01-19
Something seems wrong with the poly1305 benchmark.

Before this PR I get:
./wolfcrypt/benchmark/benchmark -poly1305
wolfCrypt Benchmark (block bytes 1048576, min 1.0 sec each)
POLY1305          6370 MB took 1.000 seconds, 6368.082 MB/s Cycles per byte =   0.48

After I get:
./wolfcrypt/benchmark/benchmark -poly1305
wolfCrypt Benchmark (block bytes 1048576, min 1.0 sec each)
POLY1305         25886900 MB took 1.000 seconds, 25886900.000 MB/s Cycles per byte =   0.00

I find that result very hard to believe.